+++ packages/article-skeleton/src/article-skeleton.js
@@ -19,7 +19,7 @@ import {
   PuzzlesWebLightTheme,
   ArticleSidebar
 } from "@times-components/ts-newskit";
-import GET_PUZZLE_SECTIONS from "@times-components/provider-queries/src/polygon-data.js";
+import fetchPolygonData from "./article-sidebar";
 import StaticContent from "./static-content";
 
 import ArticleBody, { ArticleLink } from "./article-body/article-body";
@@ -200,78 +200,15 @@ const ArticleSkeleton = ({
   const isLiveOrBreaking = getIsLiveOrBreakingFlag(expirableFlags);
   const [polygonUrl, setPolygonUrl] = useState([]);
 
-  const generatePuzzleUrl = slice => {
-    const base = "/puzzles/word-puzzles";
-
-    if (slice.shortIdentifier) {
-      return `${base}/${slice.slug}-${slice.shortIdentifier}`;
-    }
-    return base;
-  };
-
-  const fetchPolygonData = async () => {
-    try {
-      const response = await fetch("https://api.thetimes.co.uk/graphql", {
-        method: "POST",
-        body: JSON.stringify({
-          query: GET_PUZZLE_SECTIONS
-        })
-      });
-
-      if (!response.ok) {
-        throw new Error(`Network response was not ok: ${response.status}`);
-      }
-
-      const data = await response.json();
-
-      const puzzleSectionsData =
-        data && data.data && data.data.editions && data.data.editions.list
-          ? data.data.editions.list
-              .map(edition => {
-                const puzzleSection = edition.sections.find(
-                  sec => sec && sec.__typename === "PuzzleSection"
-                );
-                return puzzleSection
-                  ? {
-                      ...puzzleSection,
-                      slices: puzzleSection.slices.map(slice => ({
-                        ...slice,
-                        section: {
-                          title: puzzleSection.title,
-                          id: puzzleSection.id
-                        }
-                      }))
-                    }
-                  : null;
-              })
-              .filter(Boolean)
-          : [];
-
-      const filteredPuzzleSlices = puzzleSectionsData.flatMap(sec =>
-        sec.slices
-          .filter(
-            slice => slice.__typename === "Puzzle" && slice.type === "polygon"
-          )
-          .map(({ slug, shortIdentifier }) => ({
-            slug,
-            shortIdentifier
-          }))
-      );
-
-      const polygonData = filteredPuzzleSlices.map(puzzle =>
-        generatePuzzleUrl(puzzle)
-      );
-      setPolygonUrl(polygonData[0]);
-    } catch (error) {
-      // Handle error fetching puzzle data: log to console
-      // console.error("Error fetching puzzle data:", error);
-    }
+  const fetchPolygon = async () => {
+    const polygon = await fetchPolygonData();
+    setPolygonUrl(polygon);
   };
 
   useEffect(
     () => {
       if (CanShowPuzzleSidebar(section)) {
-        fetchPolygonData();
+        fetchPolygon();
       }
     },
     [CanShowPuzzleSidebar, section]
